backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic[email]==2.5.0

numpy==1.26.4
opencv-python==4.9.0.80

dlib==19.24.2
face-recognition==1.3.0

pymongo==4.6.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.9